refactor(agents): consolidate model normalization
