+++ src/agents/model-selection-cli.ts
@@ -4,7 +4,7 @@
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import { resolveRuntimeCliBackends } from "../plugins/cli-backends.runtime.js";
 import { resolvePluginSetupCliBackendDescriptor } from "../plugins/setup-registry.runtime.js";
-import { normalizeProviderId } from "./model-selection-normalize.js";
+import { normalizeProviderId } from "./model-ref-shared.js";
 
 /** Return true when a provider id resolves to a configured or plugin CLI backend. */
 export function isCliProvider(provider: string, cfg?: OpenClawConfig): boolean {
